Fixed warnings.
{{{#pragma omp ...}}} in scorep_trace_thread_interaction.c needs to be
replaced by SCOREP_Mutex* though.


Subversion: revision:2490

+++ src/services/metric/scorep_metric_rusage.c
@@ -276,7 +276,7 @@ scorep_metric_rusage_open()
             {
                 if ( strcmp( scorep_rusage_metrics[ i ].name, token ) == 0 )
                 {
-                    index = ( int )scorep_rusage_metrics[ i ].index;
+                    index = scorep_rusage_metrics[ i ].index;
                 }
             }
